define multiple tests for multiple scenarios
