+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-1.input.js
@@ -1,0 +1,16 @@
+/* eslint-disable */
+import { Button, Grid } from "reakit";
+
+function Component() {
+  return (
+    <div>
+      <Button as="div" />
+      <Button as={[Link, "div"]} />
+      <Grid
+        areas="a b c"
+        columns="1fr 60px auto"
+        rows="auto"
+      />
+    </div>
+  );
+}
