+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-2.input.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { as, Button } from "reakit";
+
+const ButtonDiv = as("div")(Button);
