+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-2.output.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { use, Button } from "reakit";
+
+const ButtonDiv = use(Button, "div");
