+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-3.input.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { Button } from "reakit";
+
+const ButtonDiv = Button.as("div");
