+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-4.input.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { as, Button } from "reakit";
+
+const ButtonLinkDiv = as([Link, "div"])(Button);
