+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-4.output.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { use, Button } from "reakit";
+
+const ButtonLinkDiv = use(Button, Link, "div");
