+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-5.input.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { Button } from "reakit";
+
+const ButtonLinkDiv = Button.as([Link, "div"]);
