+++ packages/reakit-codemods/src/transformations/__testfixtures__/v016-5.output.js
@@ -1,0 +1,4 @@
+/* eslint-disable */
+import { Button, use } from "reakit";
+
+const ButtonLinkDiv = use(Button, Link, "div");
