--- packages/reakit-codemods/src/transformations/__testfixtures__/v016.input.js
@@ -1,25 +1,0 @@
-/* eslint-disable */
-import { as, Button, Grid } from "reakit";
-import { Component } from "react";
-
-const ButtonDiv = as("div")(Button);
-
-const ButtonDiv = Button.as("div");
-
-const ButtonLinkDiv = as([Link, "div"])(Button);
-
-const ButtonLinkDiv = Button.as([Link, "div"]);
-
-function Component() {
-  return (
-    <div>
-      <Button as="div" />
-      <Button as={[Link, "div"]} />
-      <Grid
-        areas="a b c"
-        columns="1fr 60px auto"
-        rows="auto"
-      />
-    </div>
-  );
-}
