+++ packages/reakit-codemods/src/transformations/__tests__/v016-test.js
@@ -1,4 +1,10 @@
 jest.autoMockOff();
 const { defineTest } = require("jscodeshift/dist/testUtils");
 
-defineTest(__dirname, "v016");
+describe("V016 transformations", () => {
+  defineTest(__dirname, "v016", null, "v016-1");
+  defineTest(__dirname, "v016", null, "v016-2");
+  defineTest(__dirname, "v016", null, "v016-3");
+  defineTest(__dirname, "v016", null, "v016-4");
+  defineTest(__dirname, "v016", null, "v016-5");
+});
